# lib/observe/crate/src/types.rs
//! Shared types for the observe library.
//!
//! These are the canonical definitions. TypeScript types and JSON Schema are generated
//! from these via `bun run gen:types` (uses ts-rs + schemars).
//!
//! Generated files:
//!   shared/log-entry.generated.ts  — TypeScript interface (ts-rs)
//!   shared/schemas.json            — Combined JSON Schema: LogEntry + EntryFilter + enums
//!   shared/schemas.zod.ts          — Zod validators (gen-zod.ts reads schemas.json)
//!   shared/openapi.json            — OpenAPI 3.0 spec (gen-openapi.ts reads schemas.json)

use serde::{Deserialize, Serialize};

#[cfg(test)]
use {schemars::JsonSchema, ts_rs::TS};

// ── Log level ──────────────────────────────────────────────────────────────────

#[derive(Debug, Clone, Serialize, Deserialize)]
#[cfg_attr(test, derive(TS, JsonSchema))]
#[cfg_attr(test, ts(rename_all = "lowercase"))]
#[serde(rename_all = "lowercase")]
pub enum LogLevel { Debug, Info, Warn, Error }

// ── Log kind ───────────────────────────────────────────────────────────────────

#[derive(Debug, Clone, Serialize, Deserialize)]
#[cfg_attr(test, derive(TS, JsonSchema))]
#[cfg_attr(test, ts(rename_all = "lowercase"))]
#[serde(rename_all = "lowercase")]
pub enum LogKind { Http, App, Error }

// ── Log entry ──────────────────────────────────────────────────────────────────

#[derive(Debug, Clone, Serialize, Deserialize)]
#[cfg_attr(test, derive(TS, JsonSchema))]
#[serde(rename_all = "camelCase")]
pub struct LogEntry {
    pub ts: String,
    pub level: LogLevel,
    pub kind: LogKind,
    pub source: String,
    pub system: String,
    pub event: String,

    #[cfg_attr(test, ts(optional))] pub service:    Option<String>,
    #[cfg_attr(test, ts(optional))] pub env:        Option<String>,
    #[cfg_attr(test, ts(optional))] pub trace_id:   Option<String>,
    #[cfg_attr(test, ts(optional))] pub span_id:    Option<String>,
    #[cfg_attr(test, ts(optional))] pub request_id: Option<String>,
    #[cfg_attr(test, ts(optional))] pub error:      Option<String>,
    #[cfg_attr(test, ts(optional))] pub stack:      Option<String>,
    #[cfg_attr(test, ts(optional))] pub method:     Option<String>,
    #[cfg_attr(test, ts(optional))] pub path:       Option<String>,
    #[cfg_attr(test, ts(optional))] pub status:     Option<u16>,
    #[cfg_attr(test, ts(optional))] pub duration_ms: Option<f64>,
    #[cfg_attr(test, ts(optional))] pub cf_ray:     Option<String>,
    #[cfg_attr(test, ts(optional))] pub cf_colo:    Option<String>,
    #[cfg_attr(test, ts(optional))] pub ip:         Option<String>,
    #[cfg_attr(test, ts(optional))] pub user_agent: Option<String>,
}

// ── Entry filter ───────────────────────────────────────────────────────────────
//
// Used by the MCP query_logs tool and GET /api/debug/logs.
// All fields optional — pass only what you want to filter on.

#[derive(Debug, Clone, Serialize, Deserialize)]
#[cfg_attr(test, derive(TS, JsonSchema))]
#[serde(rename_all = "camelCase")]
pub struct EntryFilter {
    #[cfg_attr(test, ts(optional))] pub source: Option<String>,
    #[cfg_attr(test, ts(optional))] pub system: Option<String>,
    #[cfg_attr(test, ts(optional))] pub kind:   Option<LogKind>,
    #[cfg_attr(test, ts(optional))] pub level:  Option<LogLevel>,
    #[cfg_attr(test, ts(optional))] pub event:  Option<String>,
    /// ISO timestamp — return entries with ts >= since
    #[cfg_attr(test, ts(optional))] pub since:  Option<String>,
    #[cfg_attr(test, ts(optional))] pub limit:  Option<u32>,
}

// ── Binding generation (run: bun run gen:types) ────────────────────────────────

#[cfg(test)]
mod gen {
    use super::*;
    use schemars::schema_for;

    /// → shared/log-entry.generated.ts  (LogLevel, LogKind, LogEntry, EntryFilter)
    #[test]
    fn export_ts_types() {
        let content = format!(
            "// Generated by dev/gen-types.ts — do not edit.\n\
             // Source: crate/src/types.rs\n\
             // Re-generate: bun run gen:types\n\
             // Check:       bun run check:types\n\n\
             {}\n\n{}\n\n{}\n\n{}\n",
            LogLevel::decl(), LogKind::decl(), LogEntry::decl(), EntryFilter::decl(),
        );
        std::fs::write("../shared/log-entry.generated.ts", content).unwrap();
        println!("✓ shared/log-entry.generated.ts");
    }

    /// → shared/schemas.json  (combined: LogLevel, LogKind, LogEntry, EntryFilter)
    ///
    /// Consumed by:
    ///   dev/gen-zod.ts     → shared/schemas.zod.ts  (runtime validation)
    ///   dev/gen-openapi.ts → shared/openapi.json     (API documentation)
    ///   shared/mcp.ts      → MCP tool inputSchema    (AI agent tools)
    #[test]
    fn export_json_schema() {
        let entry_root  = schema_for!(LogEntry);
        let filter_root = schema_for!(EntryFilter);

        let mut defs: serde_json::Map<String, serde_json::Value> = serde_json::Map::new();
        // LogLevel + LogKind come from LogEntry's inner definitions
        for (k, v) in &entry_root.definitions {
            defs.insert(k.clone(), serde_json::to_value(v).unwrap());
        }
        defs.insert("LogEntry".into(),    serde_json::to_value(&entry_root.schema).unwrap());
        defs.insert("EntryFilter".into(), serde_json::to_value(&filter_root.schema).unwrap());

        let combined = serde_json::json!({
            "$schema": "http://json-schema.org/draft-07/schema#",
            "definitions": defs,
        });
        std::fs::write("../shared/schemas.json", serde_json::to_string_pretty(&combined).unwrap()).unwrap();
        println!("✓ shared/schemas.json");
    }
}