# lib/observe/shared/schemas.zod.ts
// Generated by dev/gen-zod.ts — do not edit.
// Source: shared/schemas.json [e149d4ff3f8f]
// Re-generate: bun run gen:zod
// Check:       bun run check:zod

import { z } from 'zod'

// ── Enums ─────────────────────────────────────────────────────────────────────

export const LogKindSchema = z.enum(["http","app","error"] as const);
export type LogKind = z.infer<typeof LogKindSchema>

export const LogLevelSchema = z.enum(["debug","info","warn","error"] as const);
export type LogLevel = z.infer<typeof LogLevelSchema>

// ── Objects ───────────────────────────────────────────────────────────────────

export const EntryFilterSchema = z.object({
  event: z.string().nullish(),
  kind: LogKindSchema.nullish(),
  level: LogLevelSchema.nullish(),
  limit: z.number().int().nullish(),
  since: z.string().nullish(),
  source: z.string().nullish(),
  system: z.string().nullish(),
}).passthrough();
export type EntryFilter = z.infer<typeof EntryFilterSchema>

export const LogEntrySchema = z.object({
  cfColo: z.string().nullish(),
  cfRay: z.string().nullish(),
  durationMs: z.number().nullish(),
  env: z.string().nullish(),
  error: z.string().nullish(),
  event: z.string(),
  ip: z.string().nullish(),
  kind: LogKindSchema,
  level: LogLevelSchema,
  method: z.string().nullish(),
  path: z.string().nullish(),
  requestId: z.string().nullish(),
  service: z.string().nullish(),
  source: z.string(),
  spanId: z.string().nullish(),
  stack: z.string().nullish(),
  status: z.number().int().nullish(),
  system: z.string(),
  traceId: z.string().nullish(),
  ts: z.string(),
  userAgent: z.string().nullish(),
}).passthrough();
export type LogEntry = z.infer<typeof LogEntrySchema>

// ── Array helpers ────────────────────────────────────────────────────────────

export const LogEntryArraySchema = z.array(LogEntrySchema)
export type LogEntryArray = z.infer<typeof LogEntryArraySchema>